requirements.txt
ipykernel
pandas
faiss-cpu
sentence-transformers
langchain-groq
python-dotenv
streamlit